tests: seperate text-decoder
